scenarioo/scenarioo#149 Moved new classes into subpackage for better structure and easier navigation.
