+++ src/main/resources/scenarioo-schemas/scenarioo-api.xsd
@@ -127,42 +127,6 @@
     </xs:sequence>
   </xs:complexType>
 
-  <xs:complexType name="screenAnnotation">
-    <xs:sequence>
-      <xs:element name="region" type="screenRegion" xmlns=""/>
-      <xs:element name="style" type="screenAnnotationStyle" minOccurs="0" xmlns=""/>
-      <xs:element name="screenText" type="xs:string" minOccurs="0"/>
-      <xs:element name="title" type="xs:string" minOccurs="0"/>
-      <xs:element name="description" type="xs:string" minOccurs="0"/>
-      <xs:element name="details">
-        <xs:complexType>
-          <xs:sequence>
-            <xs:element name="entry" minOccurs="0" maxOccurs="unbounded">
-              <xs:complexType>
-                <xs:sequence>
-                  <xs:element name="key" minOccurs="0" type="xs:string"/>
-                  <xs:element name="value" minOccurs="0" type="xs:anyType"/>
-                </xs:sequence>
-              </xs:complexType>
-            </xs:element>
-          </xs:sequence>
-        </xs:complexType>
-      </xs:element>
-      <xs:element name="clickAction" type="screenAnnotationClickAction" minOccurs="0" xmlns=""/>
-      <xs:element name="clickActionUrl" type="xs:string" minOccurs="0"/>
-      <xs:element name="clickActionText" type="xs:string" minOccurs="0"/>
-    </xs:sequence>
-  </xs:complexType>
-
-  <xs:complexType name="screenRegion">
-    <xs:sequence>
-      <xs:element name="x" type="xs:int"/>
-      <xs:element name="y" type="xs:int"/>
-      <xs:element name="width" type="xs:int"/>
-      <xs:element name="height" type="xs:int"/>
-    </xs:sequence>
-  </xs:complexType>
-
   <xs:complexType name="step">
     <xs:sequence>
       <xs:element ref="page" minOccurs="0" xmlns=""/>
@@ -226,6 +190,42 @@
           </xs:sequence>
         </xs:complexType>
       </xs:element>
+    </xs:sequence>
+  </xs:complexType>
+
+  <xs:complexType name="screenAnnotation">
+    <xs:sequence>
+      <xs:element name="region" type="screenRegion" xmlns=""/>
+      <xs:element name="style" type="screenAnnotationStyle" minOccurs="0" xmlns=""/>
+      <xs:element name="screenText" type="xs:string" minOccurs="0"/>
+      <xs:element name="title" type="xs:string" minOccurs="0"/>
+      <xs:element name="description" type="xs:string" minOccurs="0"/>
+      <xs:element name="details">
+        <xs:complexType>
+          <xs:sequence>
+            <xs:element name="entry" minOccurs="0" maxOccurs="unbounded">
+              <xs:complexType>
+                <xs:sequence>
+                  <xs:element name="key" minOccurs="0" type="xs:string"/>
+                  <xs:element name="value" minOccurs="0" type="xs:anyType"/>
+                </xs:sequence>
+              </xs:complexType>
+            </xs:element>
+          </xs:sequence>
+        </xs:complexType>
+      </xs:element>
+      <xs:element name="clickAction" type="screenAnnotationClickAction" minOccurs="0" xmlns=""/>
+      <xs:element name="clickActionUrl" type="xs:string" minOccurs="0"/>
+      <xs:element name="clickActionText" type="xs:string" minOccurs="0"/>
+    </xs:sequence>
+  </xs:complexType>
+
+  <xs:complexType name="screenRegion">
+    <xs:sequence>
+      <xs:element name="x" type="xs:int"/>
+      <xs:element name="y" type="xs:int"/>
+      <xs:element name="width" type="xs:int"/>
+      <xs:element name="height" type="xs:int"/>
     </xs:sequence>
   </xs:complexType>
 
@@ -326,6 +326,13 @@
     </xs:sequence>
   </xs:complexType>
 
+  <xs:simpleType name="status">
+    <xs:restriction base="xs:string">
+      <xs:enumeration value="SUCCESS"/>
+      <xs:enumeration value="FAILED"/>
+    </xs:restriction>
+  </xs:simpleType>
+
   <xs:simpleType name="screenAnnotationStyle">
     <xs:restriction base="xs:string">
       <xs:enumeration value="CLICK"/>
@@ -344,13 +351,6 @@
     <xs:restriction base="xs:string">
       <xs:enumeration value="TO_NEXT_STEP"/>
       <xs:enumeration value="TO_URL"/>
-    </xs:restriction>
-  </xs:simpleType>
-
-  <xs:simpleType name="status">
-    <xs:restriction base="xs:string">
-      <xs:enumeration value="SUCCESS"/>
-      <xs:enumeration value="FAILED"/>
     </xs:restriction>
   </xs:simpleType>
 </xs:schema>
