+++ src/test/java/org/scenarioo/api/ScenarioDocuWriterAndReaderTest.java
@@ -39,9 +39,6 @@ import org.scenarioo.model.docu.entities.Build;
 import org.scenarioo.model.docu.entities.Labels;
 import org.scenarioo.model.docu.entities.Page;
 import org.scenarioo.model.docu.entities.Scenario;
-import org.scenarioo.model.docu.entities.ScreenAnnotation;
-import org.scenarioo.model.docu.entities.ScreenAnnotationClickAction;
-import org.scenarioo.model.docu.entities.ScreenAnnotationStyle;
 import org.scenarioo.model.docu.entities.Step;
 import org.scenarioo.model.docu.entities.StepDescription;
 import org.scenarioo.model.docu.entities.StepHtml;
@@ -51,6 +48,9 @@ import org.scenarioo.model.docu.entities.generic.Details;
 import org.scenarioo.model.docu.entities.generic.ObjectDescription;
 import org.scenarioo.model.docu.entities.generic.ObjectList;
 import org.scenarioo.model.docu.entities.generic.ObjectReference;
+import org.scenarioo.model.docu.entities.screenAnnotations.ScreenAnnotation;
+import org.scenarioo.model.docu.entities.screenAnnotations.ScreenAnnotationClickAction;
+import org.scenarioo.model.docu.entities.screenAnnotations.ScreenAnnotationStyle;
 
 /**
  * Some smoke tests for the Scenarioo generator API.
